refactor(training): move ema to training/ema.py; add shim at ml_playground/ema.py
